Move toplevel Widget state to routableState
Instead of storing the state locally for each widget, this new technique stores
relevant state (e.g., selected wahl) within the routable state stored as an URL
parameter. This makes it possible to send a link to a layout which contains
preselected wahlids. It also improves testability of the entire UI and increases
the capability to produce more useful layout templates.

The only drawback to this approach is the decreased performance, which is acceptable however